kathak/requirements.txt
flask
flask_cors
flask_login
flask_sqlalchemy
transformers
torch
newspaper3k
Werkzeug